sd_utilities/fix_requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.9.0
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.3
attrs==23.1.0
bitsandbytes==0.38.1
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
diffusers==0.21.2
easygui==0.98.3
einops==0.6.0
entrypoints==0.4
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
google-auth==2.23.4
google-auth-oauthlib==0.4.6
grpcio==1.59.3
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
-e git+https://github.com/kohya-ss/sd-scripts.git@95ae56bd22c285ccb2fe5fca96d92f39842bb99b#egg=library
lightning-utilities==0.10.0
lit==15.0.7
Markdown==3.5.1
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
open-clip-torch==2.20.0
opencv-python==4.7.0.68
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
protobuf==3.19.6
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyre-extensions==0.0.29
python-dateutil==2.8.2
pytorch-lightning==1.9.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rpds-py==0.13.1
rsa==4.9
safetensors==0.3.1
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
timm==0.9.11
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.2.0
torchvision==0.15.2+cu118
tqdm==4.66.1
transformers==4.30.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
voluptuous==0.13.1
wcwidth==0.2.12
Werkzeug==3.0.1
xformers==0.0.20
yarl==1.9.3
zipp==3.17.0
